requirements.txt
pika
Pillow
torch
opencv-python
numpy
transformers